// src/app/models/interface/projects/response/GetAllProjectsResponse.ts
export interface GetAllProjectsResponse{
  id: number;
  name: string;
  description: string;
  startDate: string;
  endDate: string;
  status: string;
  idResponsableUser: {
    id: number;
    name: string;
    email: string;
    password: string;
    profile: string;
    creationDate: string;
    lastLogin: string;
  };
  creationDate: string;
  priority: string;
}